Adjusting getProgress()
In getProgress a current segment is ended.  I removed an uneccessary if statement below this that made a segment from the current time to the current time.  getProgress is also called in html5.js and youtube.js before determining if getArgsXAPICopmleted should be called.  getArgsXAPICompleted was also calling getProgess which results in another segment of current time to current time which is meaningless so I made progress an input to getArgsXAPICompleted so this call is only made once and the segment is ended without a meaningless one being added to the segment string.