Lowered threshold of time for endPlayedSegment
This was only pushing segments if the time was greater than 1 second.  The reason being is this would get called multiple times when a user scrubbed the play bar for a seeked event, and we used the time to determine state.  We lowered this to 1/2 a second and the results seem to go well.  Seek isn’t fired multiple times when scrubbing and play and paused are recorded acurately even with short intervals.