Revert 6dd68ac
Updating the currently playing segment start time when the video starts playing is required in certain scenarios. For example, If a user revisits a page with a youtube video that they had previously partially watched, the video will resume from where they left off. If we don’t set `playingSegmentStart` on the play event, the first playing segment recorded will start at 0 (the beginning of the video) rather than the timecode the video resumed from.