Nanootv: Use pressToPlay to hide squash overlay for login
In case of a login redirect the squash overlay prevents the user to
insert the credentials. It is also not possible to click the overlay
away, since it is only possible to do so, when setting the player state
to PLAYING. Thus the workaround with temporarily setting pressToPlay to
true works for this scenario. In order for the controls to be rendered
the pressToPlay state has to be set to false afterwards.