Add support for EchoVideos from Echo360 as source (#90)
* Add support for EchoVideos from Echo360 as source

* Clean eslint warnings

* Remove (most) jQuery

* Consistent arrow functions

* Add isLoaded function

* Iframe permission policy delegation

* JSDoc improvements and coding style

* Duration is a number or undefined

* Comparison operator

* Cancel resizing on timeline updates

* Consistent playback rate handling

* Improve start time handling

* PLT-2090: Update quality option mapping

Use updated quality options

* Fix startAt option

* Load player controls before the video starts

* JI-5798 fix auto pause when video ended (#121)

* Bump patch version

* Fix numInstances

numInstances is supposed to reflect the number of instances of the handler for a unique element id, so it cannot be an instance property

* Stop triggering "paused" before video has been "played"

Would cause trouble downstream, e.g. in Interactive Video

* Fix updating duration

When the message event is "timeline", message.data.duration is `undefined`, the duration value which is used by `getDuration()`, too, would be overwritten and be `undefined` (bug there then), and here looping would not work.

* Pass query string with iframe source

* Make eslint happy

* Make code robust for document.featurePolicy

document.featurePolicy is experimental (no support in Firefox/Safari). Prevent potential crash.

* Fix startAt handling

The parameter is passed as a float in seconds

* Add support for EchoVideos from Echo360 as source

* Clean eslint warnings

* Remove (most) jQuery

* Consistent arrow functions

* Add isLoaded function

* Iframe permission policy delegation

* JSDoc improvements and coding style

* Duration is a number or undefined

* Comparison operator

* Cancel resizing on timeline updates

* Consistent playback rate handling

* Improve start time handling

* PLT-2090: Update quality option mapping

Use updated quality options

* Fix startAt option

* Load player controls before the video starts

* Fix numInstances

numInstances is supposed to reflect the number of instances of the handler for a unique element id, so it cannot be an instance property

* Stop triggering "paused" before video has been "played"

Would cause trouble downstream, e.g. in Interactive Video

* Fix updating duration

When the message event is "timeline", message.data.duration is `undefined`, the duration value which is used by `getDuration()`, too, would be overwritten and be `undefined` (bug there then), and here looping would not work.

* Pass query string with iframe source

* Make eslint happy

* Make code robust for document.featurePolicy

document.featurePolicy is experimental (no support in Firefox/Safari). Prevent potential crash.

* Fix startAt handling

The parameter is passed as a float in seconds

* HFP-3869 fix resizing in CP editor

---------

Co-authored-by: MHod-101 <mhodson@echo360.com>
Co-authored-by: devland <5208532+devland@users.noreply.github.com>
Co-authored-by: Vilde Stabell <vilde.voss.stabell@gmail.com>
Co-authored-by: Oliver Tacke <o.tacke@posteo.de>
Co-authored-by: devland <adevland@gmail.com>